address PR review: cache surfacedFiles, extract component, Dutch label, stricter match
- Pass surfacedFiles as prop to AgentRunItem instead of computing twice
- Extract SurfacedFileCard to its own file alongside ApprovalCard
- Change "files" label to "bestanden" for Dutch UI consistency
- Use explicit toolName comparison for make_design instead of includes()
- Add inline comment explaining surfaceFileWrites config flag
